+++ Event/UserLogoutEvent.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace RevisionTen\CMS\Event;
+
+use RevisionTen\CMS\Command\UserLogoutCommand;
+use RevisionTen\CMS\Listener\UserLogoutListener;
+use RevisionTen\CQRS\Event\Event;
+use RevisionTen\CQRS\Interfaces\EventInterface;
+
+final class UserLogoutEvent extends Event implements EventInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public static function getCommandClass(): string
+    {
+        return UserLogoutCommand::class;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getListenerClass(): string
+    {
+        return UserLogoutListener::class;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getMessage(): string
+    {
+        return 'User logged out';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getCode(): int
+    {
+        return CODE_OK;
+    }
+}