+++ Handler/UserLogoutHandler.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace RevisionTen\CMS\Handler;
 
-use RevisionTen\CMS\Command\UserChangePasswordCommand;
-use RevisionTen\CMS\Event\UserChangePasswordEvent;
+use RevisionTen\CMS\Command\UserLogoutCommand;
+use RevisionTen\CMS\Event\UserLogoutEvent;
 use RevisionTen\CMS\Model\UserAggregate;
 use RevisionTen\CQRS\Interfaces\AggregateInterface;
 use RevisionTen\CQRS\Interfaces\CommandInterface;
@@ -14,7 +14,7 @@ use RevisionTen\CQRS\Interfaces\HandlerInterface;
 use RevisionTen\CQRS\Message\Message;
 use RevisionTen\CQRS\Handler\Handler;
 
-final class UserChangePasswordHandler extends Handler implements HandlerInterface
+final class UserLogoutHandler extends Handler implements HandlerInterface
 {
     /**
      * {@inheritdoc}
@@ -23,11 +23,6 @@ final class UserChangePasswordHandler extends Handler implements HandlerInterfac
      */
     public function execute(CommandInterface $command, AggregateInterface $aggregate): AggregateInterface
     {
-        $payload = $command->getPayload();
-
-        $aggregate->password = $payload['password'];
-        $aggregate->resetToken = null;
-
         return $aggregate;
     }
 
@@ -36,7 +31,7 @@ final class UserChangePasswordHandler extends Handler implements HandlerInterfac
      */
     public static function getCommandClass(): string
     {
-        return UserChangePasswordCommand::class;
+        return UserLogoutCommand::class;
     }
 
     /**
@@ -44,7 +39,7 @@ final class UserChangePasswordHandler extends Handler implements HandlerInterfac
      */
     public function createEvent(CommandInterface $command): EventInterface
     {
-        return new UserChangePasswordEvent($command);
+        return new UserLogoutEvent($command);
     }
 
     /**
@@ -52,19 +47,6 @@ final class UserChangePasswordHandler extends Handler implements HandlerInterfac
      */
     public function validateCommand(CommandInterface $command, AggregateInterface $aggregate): bool
     {
-        $payload = $command->getPayload();
-
-        if (!isset($payload['password']))  {
-            $this->messageBus->dispatch(new Message(
-                'Missing new password',
-                CODE_BAD_REQUEST,
-                $command->getUuid(),
-                $command->getAggregateUuid()
-            ));
-
-            return false;
-        }
-
         return true;
     }
 }