+++ CHANGELOG.md
@@ -10,10 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Added command to migrate old users to new user aggregates**, run `bin/console cms:user:migrate`
 - Added user password change console command
 - Added user login command/event
+- Added User password reset command
 ### Changed
 - Renamed class User to UserRead
 ### TODO
-- User password reset command (mail)
 - Create, delete & edit user in backend
 
 ## [1.2.8] - 2018-11-06